+++ attributes/default.rb
@@ -22,6 +22,7 @@
 
 default['server-base']['install_dnsmasq'] = false
 default['server-base']['install_chef_client'] = false
+default['server-base']['install_kubeadm'] = false
 
 # Python environment
 default['server-base']['python2']['install'] = true